Decrease word width from 16 bits to 8

+++ src/rom_loading.js
@@ -9,7 +9,7 @@
 var jBinary = require("jbinary");
 
 var TYPESET = {
-  "jBinary.all": ['array', 'uint16'],
+  "jBinary.all": ['array', 'uint8'],
 };
